requirements.txt
sentence-transformers
faiss-cpu
pypdf
llama-cpp-python
fastapi
uvicorn
streamlit
requests
rouge-score
